+++ sdwan_configuration_groups.tf
@@ -67,6 +67,7 @@ resource "sdwan_configuration_group" "configuration_group" {
     sdwan_policy_object_security_data_ipv4_prefix_list.policy_object_security_data_ipv4_prefix_list,
     sdwan_policy_object_security_fqdn_list.policy_object_security_fqdn_list,
     sdwan_policy_object_security_port_list.policy_object_security_port_list,
+    sdwan_policy_object_security_protocol_list.policy_object_security_protocol_list,
     sdwan_policy_object_sla_class_list.policy_object_sla_class_list
   ]
   lifecycle {